pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "genflow"
version = "0.1.2"
description = "A library for running AI workflows in the cloud."
authors = ["Matthias Georgi <matti.georgi@gmail.com>"]
packages = [
  { include = "genflow", from = "src" },
  { include = "comfy", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "*"
authlib = "*"
boto3 = ">=1.34.5"
cryptography = "*"
fastapi = ">=0.109.0"
httpx = "*"
importlib-resources = "*"
itsdangerous = "*"
joblib = "*"
jsonpath-ng = "*"
librosa = "*"
matplotlib = "*"
moto = { version = "*", extras = ["s3"] }
numpy = "*"
openai = ">=1.10.0"
openapi-pydantic = "*"
pandas = "*"
pillow = "*"
pydantic = ">=2.6.1"
pydub = "*"
pyngrok = "*"
replicate = "*"
respx = "*"
seaborn = "*"
scikit-learn = "*"
uvicorn = { version = "*", extras = ["standard"] }
validate_email = "*"
websockets = "*"
zipp = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-env = "*"
pytest-xdist = "*"


[tool.poetry.extras]
worker = [
  "accelerate",
  "addict",
  "aiohttp",
  "annotated-types",
  "av",
  "einops",
  "filelock",
  "freezegun",
  "ftfy",
  "fvcore",
  "huggingface_hub",
  "imageio",
  "importlib_metadata",
  "mediapipe",
  "modal",
  "omegaconf",
  "opencv-python-headless",
  "protobuf",
  "psutil",
  "python-dateutil",
  "pyyaml",
  "safetensors",
  "scikit-image",
  "scipy",
  "svglib",
  "torch",
  "torchsde",
  "torchvision",
  "tqdm",
  "transformers",
  "trimesh",
  "websocket-client",
  "xformers",
  "yacs",
  "yapf",
]
